Adds ToolBar.java and Algorithms.java
ToolBar.java is the GUI component that displays the
different options the user has for the path finding
algorithm, such as is diagonals are allowed or how fast
the algorithm executes. It also has the play button which
works only once. All other options need to be implemented
still.

Algorithms.java is the GUI component that displays the
different algorithms that the user can choose from. It
is currently just a GUI interface and is not functional.

Also some minor documentation and bug fixes in Grid, and
 Dijkstras .java files